+++ tests/test_read_simple_markdown.py
@@ -1,3 +1,4 @@
+import mock
 from testfixtures import compare
 import jupytext
 
@@ -102,3 +103,32 @@ def test_read_julia_notebook(markdown="""```julia
     assert nb.cells[0].cell_type == 'code'
     markdown2 = jupytext.writes(nb, ext='.md')
     compare(markdown, markdown2)
+
+
+def test_split_on_header(markdown="""A paragraph
+
+# H1 Header
+
+## H2 Header
+
+Another paragraph
+"""):
+    with mock.patch('jupytext.header.INSERT_AND_CHECK_VERSION_NUMBER', True):
+        nb = jupytext.reads(markdown, ext='.md', format_name='markdown-split-at-heading')
+        assert nb.cells[0].source == 'A paragraph'
+        assert nb.cells[1].source == '# H1 Header'
+        assert nb.cells[2].source == '## H2 Header\n\nAnother paragraph'
+        assert len(nb.cells) == 3
+        markdown2 = jupytext.writes(nb, ext='.md')
+        compare(markdown, markdown2)
+
+
+def test_split_on_header_after_two_blank_lines(markdown="""A paragraph
+
+
+# H1 Header
+"""):
+    with mock.patch('jupytext.header.INSERT_AND_CHECK_VERSION_NUMBER', True):
+        nb = jupytext.reads(markdown, ext='.Rmd', format_name='rmarkdown-split-at-heading')
+        markdown2 = jupytext.writes(nb, ext='.Rmd')
+        compare(markdown, markdown2)